requirements.txt
# -----------------------------
# Core backend framework
# -----------------------------
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0

# -----------------------------
# Environment & utilities
# -----------------------------
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0

# -----------------------------
# File upload & PDF processing
# -----------------------------
python-multipart>=0.0.9,<1.0
pypdf>=4.0,<5.0

# -----------------------------
# Database & ORM
# -----------------------------
sqlalchemy>=2.0,<3.0

# -----------------------------
# Authentication & validation
# -----------------------------
passlib[argon2]>=1.7,<2.0
python-jose[cryptography]>=3.3,<4.0
pydantic[email]>=2.5,<3.0

# -----------------------------
# Embeddings & ML
# -----------------------------
sentence-transformers>=2.6,<3.0
torch>=2.1,<3.0

# -----------------------------
# Vector database
# -----------------------------
chromadb>=0.4.24,<1.0
